rename root package to notificationService
