+++ src/main/java/com/ase/notificationService/Application.java
@@ -1,4 +1,4 @@
-package com.ase.userservice;
+package com.ase.notificationService;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
