+++ src/main/java/com/ase/notificationService/config/RepositoryConfig.java
@@ -1,4 +1,4 @@
-package com.ase.userservice.config;
+package com.ase.notificationService.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
