+++ src/main/java/com/ase/notificationService/controllers/RootController.java
@@ -1,4 +1,4 @@
-package com.ase.userservice.controllers;
+package com.ase.notificationService.controllers;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
